% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{average_residuals_over_types}
\alias{average_residuals_over_types}
\title{Sum or average residuals over types}
\usage{
average_residuals_over_types(res, weight_by_count = FALSE,
  sqrt_weights = FALSE, trim = 0)
}
\description{
Sum or average residuals over types
}
