% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-Q-spline.R, R/make-Q.R, R/make-Q2.R
\name{bdistsxy}
\alias{bdistsxy}
\title{Distance xy to bbox}
\usage{
bdistsxy(xy, bbox)

bdistsxy(xy, bbox)

bdistsxy(xy, bbox)
}
\description{
Distance xy to bbox

Distance xy to bbox

Distance xy to bbox
}
