% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV-quads.R
\name{cv_border_expected_loss}
\alias{cv_border_expected_loss}
\title{Loss of Data}
\usage{
cv_border_expected_loss(R_frac = 0.05, nx = 5)
}
\arguments{
\item{R_frac}{border range / window side length - fraction}

\item{nx}{nx x nx split}
}
\description{
How much data is wasted due to border correction in a specific nx x nx split of a square window
}
