% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV-quads.R
\name{cv_border_solve_split}
\alias{cv_border_solve_split}
\title{Solve What Kind of CV Split Can Be Done}
\usage{
cv_border_solve_split(R_frac = 0.02, acceptable_loss = 0.5)
}
\arguments{
\item{R_frac}{border range / window side length - fraction}

\item{loss}{max tolerated lost data, between (0, 1)}
}
\description{
How many quadrats can we have given border correction range and tolerated loss of data?
}
