% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-estimated-potential.R
\name{eval_estimated_potential}
\alias{eval_estimated_potential}
\title{Evaluate Estimated Potential}
\usage{
eval_estimated_potential(at, fit, k, i = 1, j = 1)
}
\arguments{
\item{fit}{a group penalised fit object}

\item{k}{which of the estimates to use}

\item{i}{species 1}

\item{j}{species 2 (can be i)}

\item{r}{range over which to evaluate the potential}
}
\description{
Evaluate Estimated Potential
}
