% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMatrix.R
\name{image.imatrix}
\alias{image.imatrix}
\title{Image for imatrix}
\usage{
image.imatrix(x, cols = c("black", "white"), zlim = c(-1, 1), ...)
}
\arguments{
\item{x}{imatrix}

\item{colors}{to use}
}
\description{
Image for imatrix
}
\details{
Colors for binary: (0,1)

Colors for signed: (-1,0,1,2)
}
