% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMatrix.R
\name{imatrix}
\alias{imatrix}
\title{Interaction Matrix for a Fitted Multivariate Multiscale Gibbs Model}
\usage{
imatrix(fit, k = NULL, group_criteria = any, signed = FALSE)
}
\arguments{
\item{fit}{A model fit}

\item{k}{which penalty level, index in lambda}

\item{group_criteria}{per interaction step-function estimate vector, how to determine overall !=0}

\item{signed}{If TRUE, m[i,j]= 0 all 0 ; +1 all ranges non-negative; -1 all non-positive; +2 mix of pos. and negative (and possibly 0).}
}
\description{
Interaction Matrix for a Fitted Multivariate Multiscale Gibbs Model
}
