% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-Q.R
\name{make_Q_stepper_multi}
\alias{make_Q_stepper_multi}
\title{Multivariate Stepper Gibbs Model Pseudolikelihood Design Matrix Generation}
\usage{
make_Q_stepper_multi(x, ranges1, ranges2, sat1 = 1, sat2 = 1,
  sat1l = NULL, sat2l = NULL, auto_sat = TRUE, covariates,
  penalise_covariates = TRUE, rho_factor = 4, rho_N = NULL,
  dum_max = 10000, dum_min = 100, dummies = NULL, save_locations = TRUE,
  ..., verb = FALSE, cov_noise_sd = 0, fatal_col0 = FALSE, border_r = 0)
}
\arguments{
\item{x}{Point pattern data, preferrably a ppp-object. Say of p-types.}

\item{ranges1}{Intra-type ranges, p vectors in a list}

\item{ranges2}{Inter-type ranges, p*(p-1)/2 vectors in a list}

\item{sat1}{Intra-type saturations, vectors (for Saturation model)}

\item{sat2}{Inter-type saturations, vectors (for Saturation model)}

\item{sat1l}{More particular sat1, matrices (for Saturation model)}

\item{sat2l}{More particular sat2, matrices (for Saturation model)}

\item{auto_sat}{Determine saturations automatically?}

\item{covariates}{list of im-objects to use as covariates}

\item{penalise_covariates}{If FALSE, covariates are estimated without penalisation.}

\item{rho_factor}{Dummy pattern intensity w.r.t. data}

\item{rho_N}{Use this many dummies (vector pf length p)}

\item{dum_max}{Upper limit for dummy count, per type (after rho_factor and rho_N)}

\item{dum_min}{Lower limit for dummy count, per type (after rho_factor and rho_N)}

\item{dummies}{Optional: (x,y,m) matrix of dummy locations to use.}

\item{save_locations}{Store the locations in the output object? Used for CV}

\item{...}{Omitted.}

\item{verb}{Verbosity}

\item{cov_noise_sd}{If > 0, add normal noise to covariates.}

\item{fatal_col0}{Should we error on zero interaction columns (singularities).}

\item{border_r}{Border correction range.}
}
\value{
A list suitable for inputting to penLogi-function
}
\description{
Construct the model matrix of multi-level multi-variate saturation model for penalised logistic regression pseudo-likelihood.
}
\details{
[todo]
}
