% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthonormalise-unorthonormalise.R
\name{orthogonalise}
\alias{orthogonalise}
\title{Orthogonalise Grouped Columns of X}
\usage{
orthogonalise(X, index)
}
\arguments{
\item{X}{matrix of coefficients (rows: coefs, cols: penalisations)}

\item{index}{grouping}
}
\description{
Orthogonalise Grouped Columns of X
}
\details{
For columns with index == 0, do nothing. For each group of columns defined by index == j,
transform the matrix such that t(Z)%*%Z/n = I, Z=X[,index==j], n = nrow(X), I = diag(1, n).
}
