% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthonormalise-unorthonormalise.R
\name{orthonormalise}
\alias{orthonormalise}
\title{Orthonormalisation for the design matrix}
\usage{
orthonormalise(X, index)
}
\arguments{
\item{X}{design matrix}

\item{index}{grouping}
}
\description{
Orthonormalisation for the design matrix
}
