% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential-details.R
\name{potential}
\alias{potential}
\title{Estimate Potential}
\usage{
potential(fit, i, j, k = NULL)
}
\arguments{
\item{fit}{Model fit returned by fitGlbin_CV}

\item{i}{Type 1 index, one of 1, ..., p}

\item{j}{Type 2 index, one of 1, ..., p}
}
\description{
Details of the estimated interaction potential between two types
}
\details{
The lasso-path coefficients of the interaction parameters between types i and j.
}
