% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{resid_by_Q}
\alias{resid_by_Q}
\title{Compute various h-residuals from group lasso fits from the Q object.}
\usage{
resid_by_Q(fit, Q, W = NULL, ...)
}
\arguments{
\item{fit}{glbinc fit}

\item{Q}{corresponding design object}

\item{W}{window for border correction}
}
\description{
The integral is estimated using MC expectation over a set of dummies.
}
