% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals_cv}
\alias{residuals_cv}
\title{Compute Spatial Residuals from CV Fits}
\usage{
residuals_cv(fit, x, dum_min = 1000, dum_max = 5 * dum_min,
  rho_factor = 10, dum_grid_dimyx = NULL, verb = FALSE, mc.cores = 1,
  weight_by_n = FALSE, ...)
}
\arguments{
\item{fit}{cv-fit}

\item{x}{original data pattern used to generate the design matrix}
}
\description{
Compute Spatial Residuals from CV Fits
}
