% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV-quads.R
\name{split_window}
\alias{split_window}
\title{Split Obs Window to Small Windows}
\usage{
split_window(bbox, W, ...)
}
\arguments{
\item{bbox}{bbox in range-column form}

\item{W}{owin-object}

\item{...}{passed to spatstat::quadrats}
}
\description{
Split Obs Window to Small Windows
}
\details{
If W not given,  try to convert bbox to owin.
}
