% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthonormalise-unorthonormalise.R
\name{standardize}
\alias{standardize}
\title{Standardize X}
\usage{
standardize(X)
}
\arguments{
\item{X}{design matrix. Will freak out if contains constant columns.}
}
\description{
Same as running 'scale' apart from result object being a list.
}
