% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthonormalise-unorthonormalise.R
\name{unorthogonalise}
\alias{unorthogonalise}
\title{Unorthonogalise Coefficients}
\usage{
unorthogonalise(beta, OG)
}
\arguments{
\item{beta}{matrix of coefficients (rows: coefs, cols: penalisations)}

\item{OG}{orthogonalisation object from 'orthogonalise'}
}
\description{
Unorthonogalise Coefficients
}
