% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthonormalise-unorthonormalise.R
\name{unorthonormalise}
\alias{unorthonormalise}
\title{Un-orthonormalisation of Coefficients}
\usage{
unorthonormalise(beta, ON)
}
\arguments{
\item{ON}{orthonormalisation setup object from 'orthonormalise'}

\item{X}{design matrix}
}
\description{
Un-orthonormalisation of Coefficients
}
