% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthonormalise-unorthonormalise.R
\name{unstandardize}
\alias{unstandardize}
\title{Unstandardise Coefficients}
\usage{
unstandardize(beta, centers, scales, cs)
}
\arguments{
\item{beta}{coef matrix (1 row per coef)}

\item{centers}{col means}

\item{scales}{col sds}

\item{cs}{alternative list with elements 'centers' and 'scales'}
}
\description{
Unstandardise Coefficients
}
